requirements.txt
ultralytics==8.3.234
opencv-python==4.12.0.88
pyzmq==26.2.0
numpy==2.2.6
protobuf
